requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
selenium==3.9.0
six==1.11.0
slackclient==1.1.2
urllib3==1.22
websocket-client==0.46.0
webwhatsapi==2.0.1
